add option to stop and remove container only

+++ web/dashboard/dasher/start_docker.sh
@@ -1,4 +1,23 @@
 set -e
+
+helpFunction()
+{
+   echo ""
+   echo "Usage: $0 [option]"
+   echo --stop "\t exits after stoping and removing container"
+   exit 1 # Exit script after printing help
+}
+
+while [ -n "$1" ]
+do
+case "$1" in
+--stop) echo "stopping and removing container only" && STOP=true;;
+*) echo "$1 is not an option" && helpFunction;;
+esac
+shift
+done
+
+# Stop and remove existing container
 IMAGE=$(docker ps -q --filter ancestor="constellation" )
 if [ "$IMAGE" ]
 then
@@ -6,6 +25,9 @@ then
 	docker stop $IMAGE
 	docker rm $IMAGE
 fi
+
+# if --stop is set, do not proceed creating new container
+if [ ! -z ${STOP+x} ]; then return 0; fi
 
 [ -d "telemetry_src" ] && rm -rf "telemetry_src"
 OLD_PATH=`pwd`
